workers/analysis/requirements.txt
# Analysis Worker Dependencies

# Google Cloud
google-cloud-bigquery>=3.11.0
google-cloud-storage>=2.10.0
google-cloud-pubsub>=2.18.0

# ML Models
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.30.0
openai-clip>=1.0.0

# Computer Vision
opencv-python>=4.8.0
pillow>=10.0.0
numpy>=1.24.0

# Audio/Video Processing
ffmpeg-python>=0.2.0
openai-whisper>=20230314

# Vector Database
pinecone-client>=2.2.0

# LLM APIs
openai>=1.0.0
anthropic>=0.3.0

# Utilities
python-dotenv>=1.0.0
pydantic>=2.0.0
